chore: Add Gruntfile, update Karma configuration & fix JSHint errors
Leveraging Grunt to check for style issues and run the Karma tests makes it
easier to run those processes. The Gruntfile in use here is based on that of
ui-utils, and the Karma config file (which has been renamed Karma.conf.js for
consistency with other projects) has been re-generated with karma init.