Update select2.js
In AngularJS 1.2.1 the newly defined controller.$render method never got called. I believe because the method already exists and the call is declared above the new declaration. Could be wrong on that though - just a guess. I just changed the method name to be unique and it is called now.